Move the collection class to its own file
